Rename module loading rules to plus names
