+++ src/webpack/rules/fontsLoader.js
